requirements.txt
autopep8==2.0.2
cffi==1.15.1
cryptography==40.0.2
greenlet==2.0.2
install==1.3.5
mysql-connector-python==8.0.33
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
protobuf==3.20.3
pycodestyle==2.10.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.13
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
